requirements.txt
# PYthon 3.11
langchain
langchain_community
langchain_pinecone
icecream
streamlit
pysqlite3-binary
langchain_openai
pinecone-client